requirements.txt
Pillow
djangorestframework
drf-yasg